Add mock bus to StartScreen for exam purposes.
The DataStoreRecord class is failing where it tries to find a bus which
is listed in the datastore but which is not contained within the
system. This is a very hacky fix solely for the purpose of the exam,
which creates the only bus used in the mock data within the DataStore
within the system on startup.

We should ideally fix the DataStoreRecord class to elegantly handle
non-existent buses, but this is a quicker solution for the exam.